deployment/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cu121
torchvision==0.16.0+cu121
pytorch-lightning>=2.0.0
gradio==3.50.2
pillow
requests
python-dotenv 
huggingface_hub>=0.19.0 